+++ retdec/decompiler.py
@@ -35,6 +35,9 @@ class Decompiler(Service):
         :param decomp_unreach_funcs: Should all functions be decompiled, even
             if they are not reachable from the main function?
         :type decomp_unreach_funcs: bool
+        :param decomp_emit_addresses: Should addresses in comments be emitted
+            in the generated code?
+        :type decomp_emit_addresses: bool
         :param architecture: Architecture. The precise meaning depends on the
             used `mode`.
         :type architecture: str
@@ -91,6 +94,7 @@ class Decompiler(Service):
         self._add_param_when_given('decomp_var_names', params, kwargs)
         self._add_param_when_given('decomp_optimizations', params, kwargs)
         self._add_param_when_given('decomp_unreach_funcs', params, kwargs)
+        self._add_param_when_given('decomp_emit_addresses', params, kwargs)
         self._add_param_when_given('architecture', params, kwargs)
         self._add_param_when_given('file_format', params, kwargs)
         self._add_param_when_given('comp_compiler', params, kwargs)