+++ tests/tools/decompiler_tests.py
@@ -408,6 +408,10 @@ class ParseArgsTests(ToolTestsBase):
         args = parse_args(['decompiler.py', '--keep-unreach-funcs', 'prog.exe'])
         self.assertTrue(args.decomp_unreach_funcs)
 
+    def test_decomp_emit_addresses_is_parsed_correctly_long_form(self):
+        args = parse_args(['decompiler.py', '--no-addresses', 'prog.exe'])
+        self.assertFalse(args.decomp_emit_addresses)
+
     def test_generate_archive_is_set_to_false_when_with_archive_not_given(self):
         args = parse_args(['decompiler.py', 'prog.exe'])
         self.assertFalse(args.generate_archive)
@@ -669,6 +673,15 @@ class MainTests(ToolTestsBase):
 
         self.assert_decompilation_was_started_also_with(
             decomp_unreach_funcs=True
+        )
+
+    def test_sets_decomp_emit_addresses_when_given(self):
+        self.call_main_with_standard_arguments_and(
+            '--no-addresses'
+        )
+
+        self.assert_decompilation_was_started_also_with(
+            decomp_emit_addresses=False
         )
 
     def test_generates_and_saves_archive_when_requested(self):